backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Database - MongoDB
mongoengine>=0.27.0
pymongo>=4.0.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Utilities
pytz>=2023.3
requests>=2.31.0
aiohttp>=3.9.0

# AI/ML (Optional - install separately if needed)
# langchain>=0.1.0
# openai>=1.0.0
google-genai>=0.8.0
# numpy>=1.24.0
# pandas>=2.0.0
